En del revision er lavet ved at ændre alle Long datatyper for Node IDer til Integer for at spare på memory. Dette vil ikke gøre alverdens forskel, men alligevel lidt fordi vi bruger flere millioner nodes. I hvert fald til OSM. Der vil ikke komme fejl i forhold til Long længere, fordi jeg har ændret NodeIDer i OSM filen til at starte fra 0 elle 1 og så opad.
Jeg har ændret en del på Dataloader klassen. Jeg har delt hele klassen op i flere metoder for at mindste om duplikationer.
Shapes og landpolygone loades nu for Krak også.

+++ Coastline/CoastlineEdge.java
@@ -9,27 +9,27 @@ import interfaces.MapEdge;
  */
 public class CoastlineEdge implements MapEdge {
     private int ID;
-    private long FNODE;
-    private long TNODE;
+    private int FNODE;
+    private int TNODE;
     private double LENGTH = 0;
     private int TYPE = 100;
     private String NAME = "";
     private String ONEWAY = "";
     private int MAX_SPEED = 0;
 
-    public CoastlineEdge(int ID, long FNODE, long TNODE) {
+    public CoastlineEdge(int ID, int FNODE, int TNODE) {
         this.ID = ID;
         this.FNODE = FNODE;
         this.TNODE = TNODE;
     }
     
     @Override
-    public long getFNode() {
+    public int getFNode() {
         return FNODE;
     }
 
     @Override
-    public long getTNode() {
+    public int getTNode() {
         return TNODE;
     }
