+++ krakkit/KrakEdgeData.java
@@ -6,8 +6,8 @@ import interfaces.MapEdge;
  * Represents the raw data from a line in kdv_unload.txt.
  */
 public class KrakEdgeData implements MapEdge {
-	public final long FNODE; 
-	public final long TNODE; 
+	public final int FNODE; 
+	public final int TNODE; 
 	public final double LENGTH;
 	public final int DAV_DK;
 	public final int DAV_DK_ID;
@@ -79,8 +79,8 @@ public class KrakEdgeData implements MapEdge {
 
 	public KrakEdgeData(String line) {
 		DataLine dl = new DataLine(line);
-		FNODE = (long) dl.getInt();
-		TNODE = (long) dl.getInt();
+		FNODE = dl.getInt();
+		TNODE = dl.getInt();
 		LENGTH = dl.getDouble();
 		DAV_DK = dl.getInt();
 		DAV_DK_ID = dl.getInt();
@@ -116,8 +116,8 @@ public class KrakEdgeData implements MapEdge {
         
         public KrakEdgeData(String line, int i){
             	DataLine dl = new DataLine(line);
-		FNODE = (long) dl.getInt();
-		TNODE = (long) dl.getInt();
+		FNODE = dl.getInt();
+		TNODE = dl.getInt();
 		LENGTH = dl.getDouble();
 		DAV_DK = 0;
 		DAV_DK_ID = 0;
@@ -152,12 +152,12 @@ public class KrakEdgeData implements MapEdge {
         }
 
     @Override
-    public long getFNode() {
+    public int getFNode() {
         return FNODE;
     }
 
     @Override
-    public long getTNode() {
+    public int getTNode() {
         return TNODE;
     }
